Extract common vector grid functionality
