+++ src/analytics/src/main/scala/osmesa/analytics/oneoffs/Footprint.scala
@@ -148,7 +148,7 @@ object Footprint extends Logging {
       .where('type === "node" and 'lat.isNotNull and 'lon.isNotNull)
       .select('lat.cast(DoubleType) as 'lat, 'lon.cast(DoubleType) as 'lon, 'key)
 
-    val stats = Footprints.createFootprints(nodes, outputURI)
+    val stats = Footprints.create(nodes, outputURI)
 
     stats.show
